+++ pkg/secrets/conjur/conjur_provider.go
@@ -1,10 +1,12 @@
-package secrets
+package conjur
 
 import (
 	"fmt"
-
 	"github.com/cyberark/conjur-api-go/conjurapi"
+	log "github.com/cyberark/conjur-authn-k8s-client/pkg/sidecar/logging"
 )
+
+var InfoLogger = log.InfoLogger
 
 func conjurProvider(tokenData []byte) (ConjurProvider, error) {
 	InfoLogger.Printf("Creating Conjur client...")
