+++ pkg/secrets/conjur/conjur_secrets_fetcher_test.go
@@ -1,20 +1,21 @@
-package secrets
+package conjur
 
 import (
+	"github.com/cyberark/conjur-authn-k8s-client/pkg/secrets"
 	. "github.com/smartystreets/goconvey/convey"
 	"reflect"
 	"testing"
 )
 
 func TestConjurSecrets(t *testing.T) {
-	Convey("GetVariableIDsToRetrieve", t, func() {
+	Convey("getVariableIDsToRetrieve", t, func() {
 		Convey("Returns true if pathMap output ia as expected", func() {
 			m := make(map[string]string)
 
 			m["account/var_path1"] = "secret1:key1"
 			m["account/var_path2"] = "secret1:key2"
 			variableIDsExpected := []string{"account/var_path1", "account/var_path2"}
-			variableIDsActual, err := GetVariableIDsToRetrieve(m)
+			variableIDsActual, err := secrets.getVariableIDsToRetrieve(m)
 
 			eq := reflect.DeepEqual(variableIDsActual, variableIDsExpected)
 
@@ -24,7 +25,7 @@ func TestConjurSecrets(t *testing.T) {
 
 		Convey("Returns error if map input is empty", func() {
 			m := make(map[string]string)
-			_, err := GetVariableIDsToRetrieve(m)
+			_, err := secrets.getVariableIDsToRetrieve(m)
 
 			So(err.Error(), ShouldEqual, "error map should not be empty")
 		})
