--- pkg/secrets/conjur_secrets_fetcher.go
@@ -1,3 +1,0 @@
-package secrets
-
-//logic of fetch sercets from Conjur
